Added -R raw_data option to allow specifying raw rdata queries.
Disallow specifying rrtype both with -t and after a slash in -n and -r options.

Disallow specifying bailiwick both with -b and after a second slash in -r options.

Renamed "type" variable to the more explicit "rrtype" variable.

Clarified in man page that the -t option specifies an rrtype, not a query type.